+++ engine/enumerate/ndoubledescription.tcc
@@ -207,7 +207,7 @@ template <class RayClass, class OutputIterator>
 void NDoubleDescription::enumerateExtremalRays(OutputIterator results,
         const NMatrixInt& subspace, const NEnumConstraintList* constraints,
         NProgressNumber* progress, unsigned initialRows) {
-    unsigned nFacets = subspace.columns();
+    size_t nFacets = subspace.columns();
 
     // If the space has dimension zero, return no results.
     if (nFacets == 0)
@@ -259,10 +259,10 @@ void NDoubleDescription::enumerateUsingBitmask(OutputIterator results,
         const NMatrixInt& subspace, const NEnumConstraintList* constraints,
         NProgressNumber* progress, unsigned initialRows) {
     // Get the dimension of the entire space in which we are working.
-    unsigned dim = subspace.columns();
+    size_t dim = subspace.columns();
 
     // Are there any hyperplanes at all in the subspace?
-    unsigned nEqns = subspace.rows();
+    size_t nEqns = subspace.rows();
     if (nEqns == 0) {
         // No!  Just send back the vertices of the non-negative orthant.
         if (progress)
